requirements.txt
# Core Biology Libraries
primer3-py
biopython

# API Framework
fastapi
uvicorn[standard]
pydantic
python-multipart

# AI/Agent
google-generativeai
python-dotenv
requests

# Data Science (optional - remove if not needed)
numpy
pandas
matplotlib
seaborn
scipy
statsmodels
scikit-learn
scikit-image
scikit-optimize